Fix three SQL bugs found by executing the migrations for the first time
I had never run this SQL. Standing it up against PostgreSQL 16 surfaced three
defects, the first of which stopped setup dead:

1. round(double precision, int) does not exist — only round(numeric, int).
   random() returns double precision, so the two lines computing a candle's
   high and low from the random walk had no matching function and aborted
   0004 partway through. Cast before rounding.

2. The guard triggers on profiles, car_orders and notifications exempted
   is_admin() but not a connection with no end user at all. The service-role
   client used by `setup:admin` and /api/admin/setup has a null auth.uid(),
   so promoting the very first administrator raised ROLE_CHANGE_FORBIDDEN and
   was impossible. Exempting a null auth.uid() is safe because RLS has already
   decided row access: RLS governs which rows, the triggers govern which
   columns a signed-in caller may change.

3. The schema relied on Supabase pre-granting table privileges to its roles —
   an unstated dependency on the platform. Applied anywhere else every query
   failed with "permission denied for table". Grants are now explicit, which
   also adds a second layer under RLS: a direct wallet write and a forged
   ledger row are now refused at the privilege layer before RLS is consulted.

Verified rather than asserted. `npm run verify:db` starts a throwaway Postgres,
applies a stand-in for the parts of Supabase the migrations touch, runs every
migration, re-runs them for idempotency, then exercises twelve behaviours: a
market buy fills at the server-held price; buys beyond cash and oversized sells
are rejected; one customer cannot see another's orders; self role escalation is
blocked; direct wallet writes, forged ledger rows and forged holdings are all
refused; the audit log is unreadable to a non-admin and immutable to everyone;
a resting limit order rests, then fills at the better price when the market
reaches it, releasing its reservation; and the service role can still bootstrap
the first administrator.

tests/migrations.test.ts adds static checks for the same classes of bug, and
fails if the round() cast is removed.

+++ supabase/migrations/0004_seed_reference_data.sql
@@ -77,8 +77,10 @@ begin
     v_drift := 0.0004;
     v_shock := (random() - 0.5) * 0.055 + (case when random() < 0.02 then (random() - 0.5) * 0.12 else 0 end);
     v_close := greatest(round((v_open * (1 + v_drift + v_shock))::numeric, 2), 5.00);
-    v_high := round(greatest(v_open, v_close) * (1 + random() * 0.018), 2);
-    v_low := round(least(v_open, v_close) * (1 - random() * 0.018), 2);
+    -- random() is double precision, and round(double precision, int) does not
+    -- exist in Postgres — only round(numeric, int). Cast before rounding.
+    v_high := round((greatest(v_open, v_close) * (1 + random() * 0.018))::numeric, 2);
+    v_low := round((least(v_open, v_close) * (1 - random() * 0.018))::numeric, 2);
     v_vol := (55_000_000 + random() * 90_000_000)::bigint;
 
     insert into public.market_candles (asset_id, interval, bucket_start, open, high, low, close, volume, is_simulated)
